Add files via upload
I added in the full file set needed in this branch because it includes the hardcoded CPU opponents. I took the code from the most recent master and made the necessary changes to the LoadBrackets() method in the MMGUI class so that the CPU's are only loaded in to ensure a 4 person scoreboard. any new players will kick a CPU out of the game.